Time: 97 ms (32.81%) | Memory: 42.9 MB (6.52%) - LeetSync

+++ 1615-range-sum-of-sorted-subarray-sums/range-sum-of-sorted-subarray-sums.cpp
@@ -3,14 +3,14 @@ public:
     int rangeSum(vector<int>& nums, int n, int left, int right) {
          vector<long long> arr;
         for (int i = 0; i < n; ++i) {
-            long sum = 0;
+            long long sum = 0;
             for (int j = i; j < n; ++j) {
                 sum += nums[j];
                 arr.push_back(sum);
             }
         }
         sort(arr.begin(), arr.end());
-        long sum = 0;
+        long long sum = 0;
         for (int l = left - 1; l < right; ++l) {
             sum += arr[l];
         }
